add some function
